@@ -2,11 +2,13 @@ import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
 import { ReactiveFormsModule } from '@angular/forms';
 import { SongUploadComponent } from './admin/song-upload/song-upload.component';
+import { AdminMessagesComponent } from './admin/admin-messages/admin-messages.component';
+import { HeaderComponent } from './header/header.component';
 
 @Component({
   selector: 'app-root',
   standalone: true,
-  imports: [ReactiveFormsModule, RouterOutlet, SongUploadComponent],
+  imports: [ReactiveFormsModule, RouterOutlet,HeaderComponent],
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css']
 })
